fixed register page and removed extra page
