+++ client/public/Asteroid.png
