+++ client/src/components/Game/Game.css
@@ -55,8 +55,7 @@
 
 .player {
   animation:3s float infinite linear;
-  transition: 2s;
-  
+  transition: 1s;
 }
 
 .player:after {
@@ -133,20 +132,8 @@ body {
 }
 
 .card-asteroid {
-  background: url("Asteroid.png");
+  background: url("/Asteroid.png");
   background-size: 100% 100%;
   background-color: transparent !important;
-  animation: 1s asteroidMove alternate infinite;
 }
 
-@keyframes asteroidMove{
-  0% {
-    
-  }
-  50% {
-
-  }
-  100% {
-
-  }
-}
