+++ client/src/signup/index.js
@@ -1,74 +1,73 @@
-import React, { Component } from "react";
+import React, { useState } from "react";
 import axios from "axios";
-import {Redirect} from "react-router-dom"
+import {Redirect} from "react-router-dom";
+import "./signup.css";
 
-class Signup extends Component {
-    state = {
-        userName: "",
-        password: ""
-      };
+function SignUp(props) {
+  const [userName, setUserName] = useState("");
+  const [passWord, setPassWord] = useState("");
+  const [error, setError] = useState("");
+
+  const updateUserName = e => {
+    setUserName(e.target.value);
+  };
+
+  const updatePassWord = e => {
+    setPassWord(e.target.value);
+  };
+
+  const submitSignUp = e => {
+    e.preventDefault();
+    axios
+      .post("/signup", {
+        username: userName,
+        password: passWord
+      })
+      .then(response => {
+        console.log(response)
+        if (response.status === 200) {
+          setError("");
+          // handle redirecting
+          props.handle(response.data);
+        } else {
+          setError("Username or Password Incorrect");
+        }
+      })
+      .catch(err => {
+        setError("UserName Already in use");
+      });
+  };
+
+  return (
     
-      handleInputChange = event => {
-        // Getting the value and name of the input which triggered the change
-        const value = event.target.value;
-        const name = event.target.name;
-    
-        // Updating the input's state
-        this.setState({
-          [name]: value
-        });
-      };
-
-    handleFormSubmit = event => {
-        // Preventing the default behavior of the form submit (which is to refresh the page)
-        event.preventDefault();
-        axios.post("/register", {
-          username: this.state.userName,
-          password: this.state.password
-        }).then(res =>{
-          console.log(res);
-        })
-        this.setState({
-            userName: "",
-            password: ""
-        });
-      };
-
-    render() {
-        return (
-          <div className="login-body">
-            { props.user !== null ? 
-              <Redirect to="/lobby" />: ""
-              }
-                    
-            <div className="card login-form">
-
-              <div className="card-body login-body">
-
-                    <h3>Register:</h3>
-                  <form className="form">
-                    <input
-                      value={this.state.userName}
-                      name="userName"
-                      onChange={this.handleInputChange}
-                      type="text"
-                      placeholder="UserName"
-                    />
-                    <input
-                      value={this.state.password}
-                      name="password"
-                      onChange={this.handleInputChange}
-                      type="text"
-                      placeholder="Password"
-                    />
-                    <button onClick={this.handleFormSubmit}>Submit</button>
-                  </form>
-              </div>
-            </div>
-          </div>
-        )
-    }
-
+    <div className="signUp-body">
+      { props.user !== null ? 
+      <Redirect to="/lobby" />: ""
+      }
+      <div className="card signUp-form">
+        <div className="card-body signUp-body">
+          <h3>Register:</h3>
+          <label htmlFor="userName">Username:</label>
+          <input
+            id="userName"
+            value={userName}
+            onChange={updateUserName}
+          ></input>
+          <label htmlFor="password">Password:</label>
+          <input
+            id="password"
+            type="password"
+            value={passWord}
+            onChange={updatePassWord}
+          ></input>
+          {error ? <p className="signUp-error">{error}</p> : ""}
+          <button className="btn btn-primary mt-1 flat btn-block" onClick={submitSignUp}>
+            Submit
+          </button>
+        </div>
+      </div>
+    </div>
+  );
 }
 
-export default Signup;
+export default SignUp;
